requirements.txt
asgiref==3.7.2
audioread==3.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0
fonttools==4.44.0
geographiclib==2.0
geopy==2.4.0
idna==3.4
imageio==2.33.0
imageio-ffmpeg==0.4.9
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
matplotlib==3.7.0
moviepy==1.0.3
msgpack==1.0.7
numba==0.58.1
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
platformdirs==3.11.0
pooch==1.8.0
proglog==0.1.10
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
